Downloadig of self-signed or invalid certificates
This change allows the download of invalid and self-signed certificates.

This removed the exception thrown for no certificate set because a server cannot start without a certificate of some sorts. Even if you run a http server on port 443 it fails earlier on a transport error since no SSL connection can be made.